edit

+++ acc/main.f90
@@ -9,7 +9,7 @@ program main
 	integer, parameter :: L = 1200
 	integer, parameter :: N = 800
 	integer, parameter :: M = 600
-	real*8 :: A(N,M), B(M,L), C(N,L)
+	integer :: A(N,M), B(M,L), C(N,L)
 	integer :: j, i, k, sum
 	real*8 :: seconds
